+++ WildFlowersReimagined/SeedMap.cs
@@ -1,9 +1,11 @@
 ﻿using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
 using StardewModdingAPI;
 using StardewValley;
 using StardewValley.GameData.Crops;
 using StardewValley.ItemTypeDefinitions;
 using System.Dynamic;
+using System.Text.RegularExpressions;
 
 namespace WildFlowersReimagined
 {
@@ -49,6 +51,23 @@ namespace WildFlowersReimagined
             checkMap.Clear();
             foreach (var (seedId, cropData) in Game1.cropData)
             {   
+                // safety check
+                if (seedId == null) {
+                    monitor.Log("Unexpected null seedId, skipping", LogLevel.Info);
+                    continue; 
+                }
+                if (cropData == null)
+                {
+                    monitor.Log($"Unexpected null cropData for seedId {seedId}, skipping", LogLevel.Info);
+                    continue;
+                }
+                if (cropData.HarvestItemId == null)
+                {
+                    monitor.Log($"Unexpected null cropData.HarvestItemId for seedId {seedId}, skipping", LogLevel.Info);
+                    continue;
+                }
+
+
                 // skip the item if it's in the seeds ignore list
                 if (ignoreList.Seeds.Contains(seedId))
                 {
@@ -145,6 +164,191 @@ namespace WildFlowersReimagined
             return seeds;
         }
     }
+
+    enum EntitlementAlgorithm
+    {
+        Equal,
+        Price,
+        Random
+    }
+
+    enum ModVanillaEntitlement
+    {
+        NoModded,
+        LessModded,
+        Normal,
+        LessVanilla,
+        NoVanilla
+    }
+
+    class EntitlementResolver
+    {
+        private const string PatternVanilla = @"^\d{3}$";
+        public bool Enabled { get; set; } = false;
+
+        public bool RespectDisabledFlowers { get; set; } = true;
+        public EntitlementAlgorithm Algorithm { get; set; } = EntitlementAlgorithm.Equal;
+        public ModVanillaEntitlement ModVanilla {  get; set; } = ModVanillaEntitlement.Normal;
+
+        public EntitlementResolver() { }
+
+        public void Resolve(IMonitor monitor, SeedMap seedMap, Dictionary<string, int> flowerProbabilityMap)
+        {
+            if (!Enabled)
+            {
+                return;
+            }
+
+            var resolveVanillaMod = this.ResolveModVanilla(seedMap);
+            var resolveAlgorithm = this.ResolveAlgorithm(seedMap);
+
+            var flowers = seedMap.GetFlowerConfigMap();
+            foreach (var (flowerInfo, SeedListData) in flowers)
+            {
+                foreach (var seedInfo in SeedListData)
+                {
+                    var seedData = seedInfo.GetParsedData();
+                    if (this.RespectDisabledFlowers && flowerProbabilityMap.GetValueOrDefault(seedData.ItemId, 3) == 0)
+                    {
+                        continue;
+                    }
+
+                    flowerProbabilityMap[seedData.ItemId] = (resolveVanillaMod.GetValueOrDefault(seedData.ItemId, 3) + resolveAlgorithm.GetValueOrDefault(seedData.ItemId, 3)) / 2;
+                }
+            }
+
+
+
+            monitor.Log("Probabilities updated", LogLevel.Info);
+            this.Enabled = false;
+        }
+
+        private Dictionary<string, int> ResolveModVanilla(SeedMap seedMap)
+        {
+            if (this.ModVanilla == ModVanillaEntitlement.Normal)
+            {
+                return new();
+            }
+            var flowers = seedMap.GetFlowerConfigMap();
+            var return_data = new Dictionary<string, int>();
+
+            var moddedValue = 3;
+            var vanillaValue = 3;
+
+            switch (this.ModVanilla) {
+                case ModVanillaEntitlement.NoModded:
+                    moddedValue = 0; 
+                    break;
+                case ModVanillaEntitlement.NoVanilla: 
+                    moddedValue = 1; 
+                    break;
+                case ModVanillaEntitlement.LessModded:
+                    moddedValue = 2;
+                    vanillaValue = 4;
+                    break;
+                case ModVanillaEntitlement.LessVanilla:
+                    moddedValue = 4;
+                    vanillaValue = 2;
+                    break;
+            }
+
+            foreach (var (flowerInfo, SeedListData) in flowers)
+            {
+                var modded = flowerInfo.QualifiedItemId.Contains('_') || !Regex.IsMatch(flowerInfo.LocalItemId, PatternVanilla);
+                var value = modded ? moddedValue : vanillaValue;
+                foreach (var seedInfo in SeedListData)
+                {
+                    var seedData = seedInfo.GetParsedData();
+                    return_data[seedData.ItemId] = value;
+                }
+
+            }
+            return return_data;
+        }
+
+        private Dictionary<string, int> ResolveAlgorithm(SeedMap seedMap) {
+            switch (this.Algorithm)
+            {
+                case EntitlementAlgorithm.Equal:
+                    return ResolveEqualAlgorithm();
+                case EntitlementAlgorithm.Price:
+                    return ResolvePriceAlgorithm(seedMap);
+                case EntitlementAlgorithm.Random:
+                    return ResolveRandomAlgorithm(seedMap);
+                default:
+                    // This path should never happen
+                    return ResolveEqualAlgorithm();
+            }
+        }
+
+        private Dictionary<string, int> ResolveEqualAlgorithm()
+        {
+            return new();
+        }
+        private Dictionary<string, int> ResolveRandomAlgorithm(SeedMap seedMap)
+        {
+            var returnValue = new Dictionary<string, int>();
+            var localRNG = new Random(Guid.NewGuid().GetHashCode());
+
+            var flowers = seedMap.GetFlowerConfigMap();
+            foreach (var (flowerInfo, SeedListData) in flowers)
+            {
+                foreach (var seedInfo in SeedListData)
+                {
+                    var seedData = seedInfo.GetParsedData();
+                    returnValue[seedData.ItemId] = localRNG.Next(0, 6);
+                }
+            }
+
+            return returnValue;
+        }
+        private Dictionary<string, int> ResolvePriceAlgorithm(SeedMap seedMap)
+        {
+            var returnValue = new Dictionary<string, int>();
+            var prices = new Dictionary<string, int>();
+
+            var flowers = seedMap.GetFlowerConfigMap();
+            int minPrice = int.MaxValue;
+            int secondMinPrice = int.MaxValue;
+            int maxPrice = int.MinValue;
+            int secondMaxPrice = int.MinValue;
+            foreach (var (flowerInfo, SeedListData) in flowers)
+            {
+                var price = flowerInfo.CreateItem(0).sellToStorePrice();
+                foreach (var seedInfo in SeedListData)
+                {
+                    var seedData = seedInfo.GetParsedData();
+                    prices[seedData.ItemId] = price;
+                }
+                if (price < minPrice){
+                    secondMinPrice = minPrice;
+                    minPrice = price;
+                }
+                if (price > maxPrice){ 
+                    secondMaxPrice = maxPrice;
+                    maxPrice = price; 
+                }
+            }
+            var quantileWidth = (secondMaxPrice - secondMinPrice) / 5;
+            foreach (var (flowerInfo, SeedListData) in flowers)
+            {
+                var price = flowerInfo.CreateItem(0).sellToStorePrice();
+                var factor = Math.Max(0, Math.Min(5, ((price - secondMinPrice) / quantileWidth)));
+                int entitlement = 5 - factor;
+                foreach (var seedInfo in SeedListData)
+                {
+                    var seedData = seedInfo.GetParsedData();
+                    returnValue[seedData.ItemId] = entitlement;
+                }
+            }
+
+            return returnValue;
+
+        }
+
+
+    }
+
 
     public sealed class IgnoreList
     {